hack around weird errors while I figure out what's happening
FIND-CHANNEL seems to be returning NIL when it shouldn't be, confusing
things.  The hack to deal with this for now is just to establish a main-loop
function to ignore errors in the same way the parser does.

Also fix more package idiocy on my part, and include a patch for
CL-IRC's PONG thing.