Fix IRC function clobbering that was causing FIND-CHANNEL errors
flybot's NICKNAME, USERNAME etc. somehow (it's apparently that I don't
understand package usage as much as I would hope) clobbered CL-IRC's
functions of the same name.  These functions happened to be accessors to
its USER type, and these returning NIL trickled down into CL-IRC not
being able to recognize its own messages (with SELF-MESSAGE-P), which
meant self-join messages not being noticed, which meant errors in
FIND-CHANNEL.  So that was fun.

The moral of the story is that I need to work out a good configuration
mechanism.