Add a cooldown timer mechanism to cut down on spam.
See cooldown.lisp for operation.

This required adding a name parameter to the timer mechanism (and
propagating this to the existing use of the timer in the vote command)
in order to distinguish command cooldown timers.

The documentation system was also generalized slightly.  See basics.lisp
for notes; it needs refactoring.